LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_boundary_events_attributes_during_drag-expected.txt


FAIL Attributes of boundary events during dragging mouse assert_equals: Pointer boundary events should have the state of primary button pressed expected "[{ type: \"pointerout\", target: <div id=\"div0\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerleave\", target: <div id=\"div0\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerover\", target: <div id=\"div1\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerenter\", target: <div id=\"div1\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerout\", target: <div id=\"div1\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerleave\", target: <div id=\"div1\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerover\", target: <div id=\"div2\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerenter\", target: <div id=\"div2\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerout\", target: <div id=\"div2\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerleave\", target: <div id=\"div2\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerover\", target: <div id=\"div3\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerenter\", target: <div id=\"div3\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerover\", target: <div id=\"div4\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerenter\", target: <div id=\"div4\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerover\", target: <div id=\"div5\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerenter\", target: <div id=\"div5\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}]" but got "[{ type: \"pointerout\", target: <div id=\"div0\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerleave\", target: <div id=\"div0\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerover\", target: <div id=\"div1\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerenter\", target: <div id=\"div1\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerout\", target: <div id=\"div1\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerleave\", target: <div id=\"div1\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerover\", target: <div id=\"div2\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerenter\", target: <div id=\"div2\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerout\", target: <div id=\"div2\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerleave\", target: <div id=\"div2\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerover\", target: <div id=\"div3\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}, { type: \"pointerenter\", target: <div id=\"div3\">, button: -1, buttons: 1, pointerId: 1, pointerType: \"mouse\"}]"
FAIL Attributes of boundary events during dragging mouse 1 assert_equals: Mouse boundary events should have the state of primary button pressed expected "[{ type: \"mouseout\", target: <div id=\"div0\">, button: 0, buttons: 1, { type: \"mouseleave\", target: <div id=\"div0\">, button: 0, buttons: 1, { type: \"mouseover\", target: <div id=\"div1\">, button: 0, buttons: 1, { type: \"mouseenter\", target: <div id=\"div1\">, button: 0, buttons: 1, { type: \"mouseout\", target: <div id=\"div1\">, button: 0, buttons: 1, { type: \"mouseleave\", target: <div id=\"div1\">, button: 0, buttons: 1, { type: \"mouseover\", target: <div id=\"div2\">, button: 0, buttons: 1, { type: \"mouseenter\", target: <div id=\"div2\">, button: 0, buttons: 1, { type: \"mouseout\", target: <div id=\"div2\">, button: 0, buttons: 1, { type: \"mouseleave\", target: <div id=\"div2\">, button: 0, buttons: 1, { type: \"mouseover\", target: <div id=\"div3\">, button: 0, buttons: 1, { type: \"mouseenter\", target: <div id=\"div3\">, button: 0, buttons: 1, { type: \"mouseover\", target: <div id=\"div4\">, button: 0, buttons: 1, { type: \"mouseenter\", target: <div id=\"div4\">, button: 0, buttons: 1, { type: \"mouseover\", target: <div id=\"div5\">, button: 0, buttons: 1, { type: \"mouseenter\", target: <div id=\"div5\">, button: 0, buttons: 1]" but got "[{ type: \"mouseout\", target: <div id=\"div0\">, button: 0, buttons: 1, { type: \"mouseleave\", target: <div id=\"div0\">, button: 0, buttons: 1, { type: \"mouseover\", target: <div id=\"div1\">, button: 0, buttons: 1, { type: \"mouseenter\", target: <div id=\"div1\">, button: 0, buttons: 1, { type: \"mouseout\", target: <div id=\"div1\">, button: 0, buttons: 1, { type: \"mouseleave\", target: <div id=\"div1\">, button: 0, buttons: 1, { type: \"mouseover\", target: <div id=\"div2\">, button: 0, buttons: 1, { type: \"mouseenter\", target: <div id=\"div2\">, button: 0, buttons: 1, { type: \"mouseout\", target: <div id=\"div2\">, button: 0, buttons: 1, { type: \"mouseleave\", target: <div id=\"div2\">, button: 0, buttons: 1, { type: \"mouseover\", target: <div id=\"div3\">, button: 0, buttons: 1, { type: \"mouseenter\", target: <div id=\"div3\">, button: 0, buttons: 1]"
